moex_volatility/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pyarrow>=8.0.0
polars>=0.16.0

# Configuration
pyyaml>=6.0

# Statistical models
statsmodels>=0.13.0
arch>=5.0.0

# Machine Learning
scikit-learn>=1.0.0
lightgbm>=3.3.0
xgboost>=1.6.0

# Hyperparameter optimization
optuna>=3.0.0

# Deep Learning (optional)
torch>=1.12.0

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.0.0

# Utilities
scipy>=1.7.0
tqdm>=4.64.0
loguru>=0.6.0
